Moving the helper methods out of test_create_preview_collection into helpers.py. Checks not run as this now generates lots of typing errors because test_* is excluded from ty, but helpers isn't
