+++ tests/e2e/deliver_grant_funding/test_add_another.py
@@ -9,6 +9,7 @@ from tests.e2e.config import EndToEndTestSecrets
 from tests.e2e.dataclasses import E2ETestUser, QuestionDict, QuestionGroupDict, QuestionResponse
 from tests.e2e.deliver_grant_funding.helpers import (
     create_grant,
+    create_question_or_group,
     extract_uuid_from_url,
     navigate_to_report_sections_page,
 )
@@ -19,7 +20,6 @@ from tests.e2e.deliver_grant_funding.reports_pages import (
     RunnerCheckYourAnswersPage,
     RunnerQuestionPage,
 )
-from tests.e2e.deliver_grant_funding.test_create_preview_collection import create_question_or_group
 from tests.e2e.helpers import delete_grant_through_admin
 
 _shared_setup_data: dict | None = None
